requirements.txt
django
psycopg2
djangorestframework-simplejwt
djangorestframework
django-cors-headers